+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Removed
+
+- The `Hive` connector has been deleted
+- The `Indexima` connector has been deleted
+- The `Rok` connector has been deleted
+- The `Lightspeed` connector has been deleted
+- The `Revinate` connector has been deleted
+
 ### [4.4.1] 2023-03-30
 
 ### Changed